### 2022-02-11 #4154 @kthyng
@ajstewartlang would this be a reasonable fit for you?
