### 2022-02-14 #4154 @ajstewartlang
@danielskatz any idea why the PDF isn't building? Looks like it's something to do with `rsvg-convert`
