### 2022-02-14 #4154 @danielskatz
From the error message, it looks like an issue with an svg image.  Perhaps the image could be stored in the different format.  If not, the help address to contact is `@openjournals/dev `
