### 2022-02-14 #4154 @ajstewartlang
@rbturnbull can you try having the image in a different format so we can see if the pdf then builds please?
