### 2022-02-14 #4154 @rbturnbull
@ajstewartlang - sure thing. I'll try the image as a pdf and see how it goes. I'm a bit surprised that the svg is causing issues because it works in github actions and using the docker image.
