### 2022-02-15 #4154 @rbturnbull
@ajstewartlang - we just changed the image from an SVG to a PDF and the paper is compiling now on https://whedon.theoj.org. Hopefully it will work for you now.
