### 2022-02-15 #4154 @rbturnbull
@ajstewartlang  - I'm so sorry, we missed adding the PDF of the actual asset in the update. I'm working on that now and I'll push that change soon and let you know.
