### 2022-02-15 #4154 @rbturnbull
@ajstewartlang the image is properly in the repository now and it is compiling fine on whedon.theoj.org. Please let me know if there are other issues with anything.
