### 2022-02-15 #4154 @ajstewartlang
Thanks @rbturnbull - the paper compiles fine now.

Also, if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
