### 2022-02-15 #4154 @rbturnbull
Thanks @ajstewartlang. I had a look at the list and I found these three potential reviewers work prefer to work in Python and are comfortable in reviewing topics in geospatial analysis:
- martinfleis
- alesarrett
- AndrewAnnex
