### 2022-02-28 #4212 @editorialbot
[REVIEW]: ausdex: A Python package for using Australian economic indexing data
**Submitting author:** @rbturnbull (<a href="http://orcid.org/0000-0003-1274-6750">Robert Turnbull</a>)
**Repository:** https://github.com/rbturnbull/ausdex
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
