### 2022-04-28 #4212 @konstantinstadler
One additional point on the article:

I think the text overpromises a bit, or perhaps there are further developments planned.

In the readme and the article you write

"
The Australian Bureau of Statistics (ABS) publishes a variety of indexes for the Australian economic environment.
These include the Consumer Price Index (CPI) used for calculating inflation and a variety of indexes designed to measure socio-economic advantage.
ausdex makes these data available in a convenient Python package with a simple programatic and command line interfaces.
"

When I first read this, my impression was that CPI is only example of the variables available through ausdex (you write "makes these data").
Are there any plans to further increase the coverage of variables by ausdex?

Similarly, you very generally describe the SEIFA module (starting at line 25). In contrast to the CPI module, the research need is not quite clear. Anyhow, based on your efforts I would guess you need the functinality for some analysis - perhaps you can describe that further in the article text.
