### 2022-04-28 #4212 @konstantinstadler
In summary, I see the need for this package and very much welcome automatic access to national statistics - many thanks for that! I do not see any major issues and hope that can be published
