### 2022-04-29 #4212 @rbturnbull
> The presented software consists of two modules: one for adjusting prices for inflation (aka "CPI module"), the other for access to the Socio-Economic Index Data for Victoria (aka "SEIFA module").
>
> Principally, I see the need for the two modules, but I am not certain they fit into the same package.
>
> The CPI module provides data for the whole of Australia which might be useful for various reasons (actually, I would use that in my own research ;-). In contrast, the SEIFA module is quite specific for one subregion of Australia - this is in odds with the general scope of the CPI module.
>
> I am not quite sure how this two packages fit together. I would suggest to either increase the scope of the SEIFA module (thus include the other areas of Australia as well). Otherwise, I would suggest to split the package in two - one specific one for CPI, the other one for SEIFA analysis of Victoria.

Thanks for your thoughts on this @konstantinstadler. We totally understand your point about the difference in scope between the two modules. We could think about splitting the packages like in one of your suggestions. In that scenario, we could imagine taking out the seifa stuff and making that its own package and leaving ausdex just to do the inflation calculation. Would it make sense then to extract all the seifa stuff from the paper submission and only have the section relevant to the inflation calculation and then we would make a new JOSS submission for the seifa package? @ajstewartlang - do you have any recommendations here?
