### 2022-05-02 9b0d49c
:abc: adding contribution guidelines as requested in issue #4
