+++ README.md
@@ -4,6 +4,7 @@
 [<img src="https://img.shields.io/endpoint?url=https://gist.githubusercontent.com/rbturnbull/49262550cc8b0fb671d46df58de213d4/raw/coverage-badge.json">](<https://rbturnbull.github.io/ausdex/coverage/>)
 [<img src="https://github.com/rbturnbull/ausdex/actions/workflows/docs.yml/badge.svg">](<https://rbturnbull.github.io/ausdex/>)
 [<img src="https://img.shields.io/badge/code%20style-black-000000.svg">](<https://github.com/psf/black>)
+[![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg)](code_of_conduct.md)
 [![status](https://joss.theoj.org/papers/817baa72d2b17b535af8f421a43404b0/status.svg)](https://joss.theoj.org/papers/817baa72d2b17b535af8f421a43404b0)
 
 An interface for several Australian socio-economic indexes.
@@ -97,23 +98,9 @@ username = {aurin_api_username}
 password = {aurin_api_password}
 ```
 
-## Development
+## Contributing
 
-To devlop ausdex, clone the repo and install the dependencies using [poetry](https://python-poetry.org/):
-
-```
-git clone https://github.com/rbturnbull/ausdex.git
-cd ausdex
-poetry install
-```
-
-You can enter the environment by running:
-
-```
-poetry shell
-```
-
-The tests can be run using `pytest`.
+See the guidelines for contributing and our code of conduct in the [documentation](https://rbturnbull.github.io/ausdex/contributing.html).
 
 ## Credits
 
