+++ pyproject.toml
@@ -39,6 +39,10 @@ pre-commit = "^2.15.0"
 sphinx-copybutton = "^0.4.0"
 black = "^21.10b0"
 
+[tool.black]
+line-length = 120
+skip_string_normalization = true
+
 [build-system]
 requires = ["setuptools >= 40.6.0", "poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
