### 2022-05-02 8db675e
:tv: adding in matplotlib as a dependency so that the examples in the docs can be run
