+++ pyproject.toml
@@ -24,6 +24,7 @@ OWSLib = "^0.25.0"
 pygeos = "^0.10.2"
 importlib-metadata = "<4.3"
 plotly = "^5.4.0"
+matplotlib = "^3.5.1"
 
 [tool.poetry.dev-dependencies]
 ipykernel = "^6.2.0"
