### 2022-05-05 #4212 @ajstewartlang
> > The presented software consists of two modules: one for adjusting prices for inflation (aka "CPI module"), the other for access to the Socio-Economic Index Data for Victoria (aka "SEIFA module").
> > Principally, I see the need for the two modules, but I am not certain they fit into the same package.
> > The CPI module provides data for the whole of Australia which might be useful for various reasons (actually, I would use that in my own research ;-). In contrast, the SEIFA module is quite specific for one subregion of Australia - this is in odds with the general scope of the CPI module.
> > I am not quite sure how this two packages fit together. I would suggest to either increase the scope of the SEIFA module (thus include the other areas of Australia as well). Otherwise, I would suggest to split the package in two - one specific one for CPI, the other one for SEIFA analysis of Victoria.
>
> Thanks for your thoughts on this @konstantinstadler. We totally understand your point about the difference in scope between the two modules. We could think about splitting the packages like in one of your suggestions. In that scenario, we could imagine taking out the seifa stuff and making that its own package and leaving ausdex just to do the inflation calculation. Would it make sense then to extract all the seifa stuff from the paper submission and only have the section relevant to the inflation calculation and then we would make a new JOSS submission for the seifa package? @ajstewartlang - do you have any recommendations here?

@rbturnbull just to summarise - option 1 is increasing the geographic scope of the seifa module vs option 2 which is removing the seifa content and putting that in a separate submission.  If increasing the scope of the seifa module was easy enough to do, this would probably be the way to go - mainly because it would increase the contribution of the present submission. But if this approach isn't straightforward then splitting the seifa package off into a separate submission might be ok (although it will then have quite a limited scope if it's just applied to one region - and I'd suggest it should probably be broader than just Victoria) - this will also reduce the lines of code in the present submission by (I think) just over 1000. I have a bias towards option 1 (which means both modules apply to the same geographic region), but option 2 looks maybe ok as well if you want to go down that route (although I do wonder where a Victoria-focused seifa module would really be enough for a submission in its own right - I suspect you'll want to broaden this geographic scope either way).
