### 2022-05-10 #4212 @rbturnbull
Thanks @ajstewartlang and @konstantinstadler . Jon and I think that we will split the two modules. We will take out the SEIFA module from ausdex and keep ausdex focussed on things related to CPI. We will edit the paper text to reflect that. Then we will expand the scope of SEIFA and then make a separate submission for that later on. Does that sound OK?
