### 2022-05-10 cfc3547
:zap: removing seifa module. This will be at https://github.com/sailngarbwm/seifa-gis
