+++ paper.md
@@ -4,7 +4,6 @@ tags:
   - Python
   - economics
   - inflation
-  - socio-economic advantage
 authors:
   - name: Robert Turnbull^[co-first author] # note this makes a footnote saying 'co-first author'
     orcid: 0000-0003-1274-6750
@@ -15,7 +14,7 @@ authors:
 affiliations:
  - name: Melbourne Data Analytics Platform, University of Melbourne
    index: 1
-date: 4 October 2021
+date: 10 May 2021
 bibliography: paper.bib
 
 ---
@@ -30,78 +29,8 @@ available in a convenient Python package with a simple programatic and command-l
 
 # Statement of need
 
-`ausdex` is a Python package for querying data produced by the ABS and returning them in a convenient format. Currently ABS data is typically housed in Microsoft Excel spreadsheets linked from the data catalogue. This package interfaces with a subset of the the data to provide an Application Programming Interface (API) to derived economic metrics. For example, we expose the Australian consumer price index data to create an inflation calculator similar to the [cpi](https://github.com/palewire/cpi) Python package for adjusting US dollars. In addition, we bring API access to ABS Socio-Economic Index Data for Areas (SEIFA) aggregated at the suburb level in Victoria. This allows for quick assessment of the socio-economic history of a suburb in Victoria from historical census data. These datasets are housed in different online repositories and are aggregated to different spatial extents since statistical geographic boundaries are redrawn from every census dataset. The `ausdex.seifa_vic` submodule eliminates the need for downloading and combining datasets from different data sources, and allows for time series comparisons tied to the current suburb geographic boundaries. 
+`ausdex` is a Python package for querying data produced by the ABS and returning them in a convenient format. Currently ABS data is typically housed in Microsoft Excel spreadsheets linked from the data catalogue. This package interfaces with a subset of the the data to provide an Application Programming Interface (API) to derived economic metrics. For example, we expose the Australian consumer price index data to create an inflation calculator similar to the [cpi](https://github.com/palewire/cpi) Python package for adjusting US dollars.
 
-# Socio-economic indexes aggregated from census data for Victoria
-Since 1986 (see table 1), the ABS has generated "Socio-Economic Indexes For Areas" (SEIFA) following each census [@seifa2016]. These indexes are aggregations of socio-economic inputs from the census forms (i.e. household income, rental/mortgage price, educational level) at the "census district level" or "mesh level" (2006–current). Census districts, or mesh levels, are geographic areas statistically defined from the census data to be the largest scale (smallest) geographic building blocks of demographic and socio-economic data based on population distribution. These statistical geographies are redrawn after each census. The ABS aggregates these to other statistical "levels" of geographic area from the Australian Statistical Geography Standard (ASGS) ([Statistical Areas Levels 1–4](https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026)) suburbs [@abs_2016_sa1_shape] and local government areas in their "Data Cube" outputs. 
-
-<!-- 
[…]
