+++ paper.bib
@@ -95,150 +95,6 @@
   url = {https://www.abs.gov.au/ausstats/abs@.nsf/mf/6461.0},
   urldate = {2021-11-30}
 }
-@online{seifa2016,
-  author = {{Australian Bureau of Statistics}},
-  title = {2033.0.55.001 - {Census of Population and Housing: Socio-Economic Indexes for Areas (SEIFA), Australia}},
-  year = 2016,
-  url = {https://www.abs.gov.au/ausstats/abs@.nsf/mf/2033.0.55.001},
-  urldate = {2021-11-30}
-}
-
-@article{aurin_portal,
-author = {Sinnott, Richard O. and Bayliss, Christopher and Bromage, Andrew and Galang, Gerson and Grazioli, Guido and Greenwood, Philip and Macaulay, Angus and Morandini, Luca and Nogoorani, Ghazal and Nino-Ruiz, Marcos and Tomko, Martin and Pettit, Christopher and Sarwar, Muhammad and Stimson, Robert and Voorsluys, William and Widjaja, Ivo},
-title = {{The Australia urban research gateway}},
-journal = {Concurrency and Computation: Practice and Experience},
-volume = {27},
-number = {2},
-pages = {358-375},
-keywords = {urban research, e-Social Science, e-Health, e-Planning},
-doi = {10.1002/cpe.3282},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cpe.3282},
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/cpe.3282},
-abstract = {Summary The \$20m Australian Urban Research Infrastructure Network (AURIN) project (www.aurin.org.au) began in July 2010. AURIN has been tasked with developing a secure, Web-based virtual environment (e-Infrastructure) offering seamless, secure access to diverse, distributed and extremely heterogeneous data sets from numerous agencies with an extensive portfolio of targeted analytical and visualization tools. This is being provisioned for Australia-wide urban and built environment researchers – itself a highly heterogeneous collection of research communities with diverse demands, through a unified urban research gateway. This paper describes these demands and how the e-Infrastructure and gateway is being designed and implemented to accommodate this diversity of requirements, both from the user/researcher perspective and from the data provider perspective. The scaling of the infrastructure is presented and the way in which it copes with the spectrum of big data challenges (volume, veracity, variability and velocity) and associated big data analytics. The utility of the e-Infrastructure is also demonstrated through a range of scenarios illustrating and reflecting the interdisciplinary urban research now possible. Copyright © 2014 John Wiley \& Sons, Ltd.},
-year = {2015}
-}
-
-@misc{abs_2006_cd_shape,
-  title = {Collection District Digital Boundaries, Victoria (ASGC 2006) in ESRI Shapefile Format  },
-  author = {{Australian Bureau of Statistics}},
-  year = {2006},
-  howpublished = {\href{https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/1259.0.30.0022006?OpenDocument}
-                        {Statistical Geography - Australian Standard Geographical Classification (ASGC), Digital Boundaries}},
-  note = {Accessed: 2021-11}
-}
-
-@misc{abs_2011_sa1_shape,
[…]
