+++ make_test_config.sh
@@ -1,4 +1,0 @@
-touch config.ini
-echo "[aurin]" >> config.ini
-echo "username = $USERNAME" >>config.ini
-echo "password = $PASSWORD" >>config.ini

[…]
