### 2022-05-10 2cdc998
:zap: fixing date to decimal year calculation to make it more accurate and updating tests. 100% coverage now
