+++ tests/test_dates.py
@@ -19,14 +19,10 @@ class TestDates(unittest.TestCase):
         np.testing.assert_equal(result, np.array("2006-01-01", dtype="datetime64[D]"))
 
     def test_int(self):
-        np.testing.assert_equal(
-            convert_date(2006), np.array("2006-01-01", dtype="datetime64[D]")
-        )
+        np.testing.assert_equal(convert_date(2006), np.array("2006-01-01", dtype="datetime64[D]"))
 
     def test_float(self):
-        np.testing.assert_equal(
-            convert_date(2006.5), np.array("2006-07-02", dtype="datetime64[D]")
-        )
+        np.testing.assert_equal(convert_date(2006.5), np.array("2006-07-02", dtype="datetime64[D]"))
 
     def test_datetime(self):
         np.testing.assert_equal(
@@ -74,6 +70,22 @@ class TestDateToDecimalYear(unittest.TestCase):
         self.assertIsInstance(result, np.ndarray)
         self.assertEqual(result, 1995.5)
 
+    def test_float_list(self):
+        result = date_time_to_decimal_year([1995.5])
+        self.assertIsInstance(result, np.ndarray)
+        self.assertEqual(result, 1995.5)
+
+    def test_datetime(self):
+        result = date_time_to_decimal_year(datetime(1996, 7, 2))
+        self.assertIsInstance(result, np.ndarray)
+        self.assertEqual(result, 1996.5)
+
+    def test_float_tuple(self):
+        result = date_time_to_decimal_year((1995.5, 1996.6))
+        self.assertIsInstance(result, np.ndarray)
+        self.assertEqual(result[0], 1995.5)
+        self.assertEqual(result[1], 1996.6)
+
     def test_float_array(self):
         result = date_time_to_decimal_year(np.array([1995.5, 1996.6]))
         self.assertIsInstance(result, np.ndarray)
@@ -85,57 +97,47 @@ class TestDateToDecimalYear(unittest.TestCase):
         self.assertTrue(all([x == y for x, y in zip(result, [1995, 1996])]))
 
     def test_datetime_array(self):
-        result = date_time_to_decimal_year(
-            np.array([np.datetime64("1995-01-01"), np.datetime64("1996-06-30")])
-        )
+        result = date_time_to_decimal_year(np.array([np.datetime64("1995-01-01"), np.datetime64("1996-07-02")]))
         self.assertIsInstance(result, np.ndarray)
         # trues = [x == y for x, y in zip(result, [1995.5, 1996.6])]
-        self.assertAlmostEqual(result[0], 1995.0833, 4)
+        self.assertAlmostEqual(result[0], 1995.0, 4)
         self.assertEqual(result[1], 1996.5)
 
     def test_str_array(self):
-        result = date_time_to_decimal_year(np.array(["1995-01-01", "1996-06-30"]))
+        result = date_time_to_decimal_year(np.array(["1995-01-01", "1996-07-02"]))
         self.assertIsInstance(result, np.ndarray)
         # trues = [x == y for x, y in zip(result, [1995.5, 1996.6])]
-        self.assertAlmostEqual(result[0], 1995.0833, 4)
+        self.assertAlmostEqual(result[0], 1995.0, 4)
         self.assertEqual(result[1], 1996.5)
 
     def test_str_series(self):
-        result = date_time_to_decimal_year(
-            pd.Series(np.array(["1995-01-01", "1996-06-30"]))
-        )
+        result = date_time_to_decimal_year(pd.Series(np.array(["1995-01-01", "1996-07-02"])))
         self.assertIsInstance(result, np.ndarray)
         # trues = [x == y for x, y in zip(result, [1995.5, 1996.6])]
[…]
