+++ _sources/reference.rst.txt
@@ -1,16 +1,6 @@
-
 =======================
 API Reference
 =======================
-
-Socio-economic Indices 
-======================
-
-.. automodule:: ausdex.seifa_vic.seifa_vic
-   :members:
-
-Inflation 
-=========
 
 .. automodule:: ausdex.inflation
    :members:   
