+++ .github/workflows/coverage.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [3.9]
+        python-version: ["3.7", "3.8", "3.9", "3.10"]
 
     steps:
       
@@ -23,12 +23,6 @@ jobs:
         python -m pip install --upgrade pip
         pip install poetry
         poetry install
-    - name: Create config file for aurin
-      run: |
-        bash make_test_config.sh
-      env:
-        USERNAME: ${{ secrets.AURINUSERNAME }}
-        PASSWORD: ${{ secrets.AURINPASSWORD }}
     - name: Coverage
       run: |
         poetry run coverage run -m pytest
