### 2022-05-10 e73b363
:abc: updating docs with CLI reference
