+++ docs/conf.py
@@ -40,6 +40,7 @@ extensions = [
     "sphinx.ext.coverage",
     "sphinx.ext.napoleon",
     "sphinx_copybutton",
+    "sphinx_click",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
