+++ pyproject.toml
@@ -16,10 +16,11 @@ openpyxl = "^3.0.0"
 xlrd = "^2.0.1"
 cached-property = "^1.5.2"
 python-dateutil = "^2.8.2"
-typer = "^0.3.2"
+typer = "^0.4.0"
 modin = "^0.10.2"
 importlib-metadata = "<4.3"
 plotly = "^5.4.0"
+kaleido = "0.2.1"
 
 [tool.poetry.dev-dependencies]
 ipykernel = "^6.2.0"
@@ -34,6 +35,7 @@ myst-parser = "^0.15.2"
 pre-commit = "^2.15.0"
 sphinx-copybutton = "^0.4.0"
 black = "^21.10b0"
+sphinx-click = {git = "https://github.com/rbturnbull/sphinx-click.git"}
 
 [tool.black]
 line-length = 120
