+++ _sources/cli.rst.txt
@@ -1,0 +1,7 @@
+============================================
+Command-line Interface Reference
+============================================
+
+.. click:: ausdex.main:app
+   :prog: ausdex
+   :nested: full
