+++ _sources/index.rst.txt
@@ -22,6 +22,7 @@ available in a convenient Python package with a simple programatic and command l
 
    quickstart
    reference
+   cli
    citation
    contributing
    
