+++ _sources/installation.md.txt
@@ -1,25 +1,0 @@
-# Installation
-
-You can install `ausdex` from the Python Package Index (PyPI):
-
-```
-pip install ausdex
-```
-
-## Development
-
-To devlop ausdex, clone the repo and install the dependencies using [poetry](https://python-poetry.org/):
-
-```
-git clone https://github.com/rbturnbull/ausdex.git
-cd ausdex
-poetry install
-```
-
-You can enter the environment by running:
-
-```
-poetry shell
-```
-
-The tests can be run using `pytest`.
