### 2022-05-10 6e1b44d
:abc: adding help for args for plot_inflation
