### 2022-05-11 e7e72f7
:abc: making disclaimer of license explicit in readme and docs
