+++ ausdex/inflation.py
@@ -13,7 +13,7 @@ import plotly.express as px
 
 from cached_property import cached_property
 
-from .files import cached_download, get_cached_path
+from .files import cached_download, get_cached_path, DownloadError
 from .dates import convert_date
 
 
@@ -40,15 +40,20 @@ class CPI:
         if quarter not in self.ACCEPTED_QUARTERS:
             raise ValueError(f"Cannot understand quarter {quarter}.")
 
-        url = f"https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia/{quarter}-{year}/{id}.xls"
+        try:
+            url = f"https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia/{quarter}-{year}/{id}.xlsx"
+            local_path = get_cached_path(f"{id}-{quarter}-{year}.xlsx")
+            cached_download(url, local_path)
+        except (DownloadError, IOError):
+            url = f"https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia/{quarter}-{year}/{id}.xls"
+            local_path = get_cached_path(f"{id}-{quarter}-{year}.xls")
+            cached_download(url, local_path)
 
-        local_path = get_cached_path(f"{id}-{quarter}-{year}.xls")
-
-        cached_download(url, local_path)
         return local_path
 
     def get_abs_by_date(self, id: str, date: datetime) -> Path:
-        """Gets a datafile from the Australian Burau of Statistics before a specific date.
+        """
+        Gets a datafile from the Australian Burau of Statistics before a specific date.
 
         Args:
             id (str): The ABS id for the datafile. For Australian Consumer Price Index (CPI) the ID is 640101.
@@ -76,7 +81,8 @@ class CPI:
         return file
 
     def latest_cpi_datafile(self) -> Path:
-        """Returns the path to the latest cached file with the Australian Consumer Price Index (CPI) data.
+        """
+        Returns the path to the latest cached file with the Australian Consumer Price Index (CPI) data.
 
         The ABS id of this file is "640101".
 
@@ -87,7 +93,8 @@ class CPI:
 
     @cached_property
     def latest_cpi_df(self) -> pd.DataFrame:
-        """Returns a Pandas DataFrame with the latest Australian Consumer Price Index (CPI) data.
+        """
+        Returns a Pandas DataFrame with the latest Australian Consumer Price Index (CPI) data.
 
         Returns:
             pd.DataFrame: The latest Australian Consumer Price Index (CPI) data. The index of the series is the relevant date for each row.
@@ -106,7 +113,8 @@ class CPI:
 
     @cached_property
     def cpi_australia_series(self) -> pd.Series:
-        """Returns a Pandas Series with the Australian CPI (Consumer Price Index) per quarter.
+        """
+        Returns a Pandas Series with the Australian CPI (Consumer Price Index) per quarter.
 
         This is taken from the latest spreadsheet from the Australian Bureau of Statistics (file id '640101'). The index of the series is the relevant date.
 
@@ -117,7 +125,8 @@ class CPI:
         return df["Index Numbers ;  All groups CPI ;  Australia ;"]
 
     def cpi_australia_at(self, date: Union[datetime, str, pd.Series, np.ndarray]) -> Union[float, np.ndarray]:
-        """Returns the CPI (Consumer Price Index) for a date (or a number of dates).
+        """
+        Returns the CPI (Consumer Price Index) for a date (or a number of dates).
 
         If `date` is a string then it is converted to a datetime using dateutil.parser.
         If `date` is a vector then it returns a vector otherwise it returns a single scalar value.
@@ -149,7 +158,8 @@ class CPI:
         original_date: Union[datetime, str],
[…]
