### 2022-05-11 5d72212
:zap: adding way to get the CPI for a particular location
