+++ ausdex/inflation.py
@@ -1,11 +1,11 @@
 import sys
+from enum import Enum
 from pathlib import Path
 from datetime import datetime, timedelta
 from typing import Union
 import pandas as pd
 from pandas._config import config
 import modin.pandas as mpd
-from dateutil import parser
 import numpy as np
 import numbers
 import plotly
@@ -15,6 +15,22 @@ from cached_property import cached_property
 
 from .files import cached_download, get_cached_path, DownloadError
 from .dates import convert_date
+from .viz import format_fig
+
+
+class Location(str, Enum):
+    AUSTRALIA = "Australia"
+    SYDNEY = "Sydney"
+    MELBOURNE = "Melbourne"
+    BRISBANE = "Brisbane"
+    ADELAIDE = "Adelaide"
+    PERTH = "Perth"
+    HOBART = "Hobart"
+    DARWIN = "Darwin"
+    CANBERRA = "Canberra"
+
+    def __str__(self):
+        return self.value
 
 
 class CPI:
@@ -40,14 +56,16 @@ class CPI:
         if quarter not in self.ACCEPTED_QUARTERS:
             raise ValueError(f"Cannot understand quarter {quarter}.")
 
+        def cached_download_abs(quarter, year, id, extension):
+            url = f"https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia/{quarter}-{year}/{id}.{extension}"
+            local_path = get_cached_path(f"{id}-{quarter}-{year}.{extension}")
+            cached_download(url, local_path)
+            return local_path
+
         try:
-            url = f"https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia/{quarter}-{year}/{id}.xlsx"
-            local_path = get_cached_path(f"{id}-{quarter}-{year}.xlsx")
-            cached_download(url, local_path)
+            local_path = cached_download_abs(quarter=quarter, year=year, id=id, extension="xlsx")
         except (DownloadError, IOError):
-            url = f"https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia/{quarter}-{year}/{id}.xls"
-            local_path = get_cached_path(f"{id}-{quarter}-{year}.xls")
-            cached_download(url, local_path)
+            local_path = cached_download_abs(quarter=quarter, year=year, id=id, extension="xls")
 
         return local_path
 
@@ -111,20 +129,28 @@ class CPI:
 
         return df
 
-    @cached_property
-    def cpi_australia_series(self) -> pd.Series:
+    def cpi_series(self, location: Union[Location, str] = Location.AUSTRALIA) -> pd.Series:
         """
         Returns a Pandas Series with the Australian CPI (Consumer Price Index) per quarter.
 
-        This is taken from the latest spreadsheet from the Australian Bureau of Statistics (file id '640101'). The index of the series is the relevant date.
+        This is taken from the latest spreadsheet from the Australian Bureau of Statistics (file id '640101').
+        The index of the series is the relevant date.
+
+        Args:
+            location (Union[Location, str], optional): The location for calculating the CPI.
+                Options are 'Australia', 'Sydney', 'Melbourne', 'Brisbane', 'Adelaide', 'Perth', 'Hobart', 'Darwin', and 'Canberra'.
+                Default is 'Australia'.
 
         Returns:
             pd.Series: The CPIs per quarter.
         """
         df = self.latest_cpi_df
-        return df["Index Numbers ;  All groups CPI ;  Australia ;"]
 
-    def cpi_australia_at(self, date: Union[datetime, str, pd.Series, np.ndarray]) -> Union[float, np.ndarray]:
+        return df[f"Index Numbers ;  All groups CPI ;  {location} ;"]
+
+    def cpi_at(
[…]
