### 2022-05-11 cca83bf
:cop: adding tests for other locations
