### 2022-05-11 4fa265e
:zap: getting CPI plot to work with locations
