### 2022-05-11 e3f9b12
:zap: making the plot inflation function more versatile
