### 2022-05-11 3ff4f5e
:abc: explaining in the docs that different locations can be given
