### 2022-05-11 04f34ce
:abc: updating readme with an example of using location
