### 2022-05-11 eba8d8d
:bug: removing matplotlib that came from vscode autoimport :(
