### 2022-05-11 224119c
:cop: moving tests for plotting CLI into test_main.py
