+++ tests/test_inflation.py
@@ -160,40 +160,3 @@ class TestInflationPlot(unittest.TestCase):
                 self.tmp / "test_cpi_fig.json",
             )
         )
-
-    def test_inflation_graph_cli(self):
-
-        from typer.testing import CliRunner
-        from ausdex import main
-
-        runner = CliRunner()
-
-        runner.invoke(
-            main.app,
-            [
-                "plot-inflation",
-                "01-01-2019",
-                str(self.tmp / "test_inflation_plot.html"),
-                "--start-date",
-                "06-06-1949",
-            ],
-        )
-        assert (self.tmp / "test_inflation_plot.html").exists()
-
-    def test_cpi_graph_cli(self):
-
-        from typer.testing import CliRunner
-        from ausdex import main
-
-        runner = CliRunner()
-
-        runner.invoke(
-            main.app,
-            [
-                "plot-cpi",
-                str(self.tmp / "test_cpi_plot.png"),
-                "--start-date",
-                "06-06-1949",
-            ],
-        )
-        assert (self.tmp / "test_cpi_plot.png").exists()
