+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'ausdex: A Python package for using Australian economic indexing data'
+title: 'ausdex: A Python package for adjusting Australian dollars for inflation'
 tags:
   - Python
   - economics
@@ -21,41 +21,38 @@ bibliography: paper.bib
 
 # Summary
 
-The Australian Bureau of Statistics (ABS) publishes a variety of indexes for the Australian
-economic environment. These include the Consumer Price Index (CPI) used for calculating inflation
-and various indexes designed to measure socio-economic advantage. `ausdex` makes these data
-available in a convenient Python package with a simple programatic and command-line interfaces. 
-
+The Australian Bureau of Statistics (ABS) publishes the Consumer Price Index (CPI) 
+for Australia and its capital cities which allows for adjustment of the value of Australia dollars for inflation. 
+`ausdex` makes these data available with an inflation calculator
+in a convenient Python package with simple programatic and command-line interfaces.
 
 # Statement of need
 
-`ausdex` is a Python package for querying data produced by the ABS and returning them in a convenient format. Currently ABS data is typically housed in Microsoft Excel spreadsheets linked from the data catalogue. This package interfaces with a subset of the the data to provide an Application Programming Interface (API) to derived economic metrics. For example, we expose the Australian consumer price index data to create an inflation calculator similar to the [cpi](https://github.com/palewire/cpi) Python package for adjusting US dollars.
+The ABS data is generally housed in Microsoft Excel spreadsheets linked from the data catalogue. Working with these spreadsheets directly is cumbersome. This package provides an Application Programming Interface (API) for Australian CPI data that seemlessly interoperates with `NumPy` [@harris2020array] and `pandas` [@mckinney-proc-scipy-2010]. It makes working with Australian dollars in Python convenient in a similar manner to the [cpi](https://github.com/palewire/cpi) Python package which adjusts US dollars for inflation.
 
-
-# Inflation
-
-![Figure 1](docs/images/cpi-time.pdf)
-<p align = "center"> Figure 1: The CPI in Australia since 1948.</p>
+# Calculating Inflation
 
 The Consumer Price Index (CPI) is a weighted average price of a basket of goods and services for urban consumers [@ABS_CPI_Methods]. The ABS issues the Australian CPI [each quarter](https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia). The data are available from September 1948 onwards (see fig. 1). The CPI values before decimalization of the Australian currency on February 14, 1966 are in understood according to the conversion rates specified in the 1965 Year Book of Australia such that £1 is equivalent to $2 [@yearbook1965, p. 810].
+
+![The CPI in Australia since 1948](docs/images/cpi-time.pdf)
 
 To adjust prices for inflation, we assume that the ratio of prices for two dates is equal to ratio of the CPIs for those dates [@parkin_macroeconomics_2019, p. 811]. This gives the formula:
 
 $$ \textrm{Price at time B} = \textrm{Price at time A} \times \frac{\textrm{CPI at time B}}{\textrm{CPI at time A}} $$
 
-The `ausdex` package automates the process for downloading the latest version of the Australian CPI data from the ABS. The user enters a price, the original date ($A$) and the evaluation date ($B$) and it returns the adjusted price. The inputs can be scalar values or vectors as a `NumPy` array [@harris2020array], a `pandas` series [@mckinney-proc-scipy-2010] or a `Modin pandas` series [@Petersohn2020]. The size of the returned vector of prices is the same as the vector of original prices. If the original date or the evaluation dates are vectors instead of scalar values then these must be the same size as the vector of prices.
+# Package Description
 
[…]
