+++ ausdex/main.py
@@ -147,6 +147,7 @@ def plot_cpi(
         help="The location for calculating the CPI.",
         case_sensitive=False,
     ),
+    title: str = typer.Option(None, help="A custom title of the plot."),
 ):
     """
     Plot the Australian CPI over time.
@@ -155,10 +156,12 @@ def plot_cpi(
         output (Path): The path to where the figure will be saved. Output can be PDF, SVG, JPG, PNG or HTML based on the extension.
         start_date (str, optional): Date to set the beginning of the time series graph. If empty, it defaults to 1948.
         end_date (str, optional): Date to set the end of the time series graph too. If empty, then the end date to the most recent quarter.
+        location (List[location]): The location for calculating the CPI.
+        title (str, optional): A custom title of the plot.
     """
     from ausdex.inflation import plot_cpi_timeseries
 
-    fig = plot_cpi_timeseries(start_date=start_date, end_date=end_date, locations=location)
+    fig = plot_cpi_timeseries(start_date=start_date, end_date=end_date, locations=location, title=title)
     if output:
         print(f"Writing figure to '{output}'.")
         viz.write_fig(fig, output)
