+++ paper.bib
@@ -114,6 +114,16 @@ pages = {2033–2046},
 numpages = {14}
 }
 
+@article{hung2021output,
+  title={Output-inflation trade-off in the presence of foreign capital: Evidence for Vietnam},
+  author={Hung, Ly Dai},
+  journal={South Asian Journal of Macroeconomics and Public Finance},
+  volume={10},
+  number={2},
+  pages={179--192},
+  year={2021},
+  publisher={SAGE Publications Sage India: New Delhi, India}
+}
 
 
 
