+++ paper.md
@@ -22,25 +22,25 @@ bibliography: paper.bib
 # Summary
 
 The Australian Bureau of Statistics (ABS) publishes the Consumer Price Index (CPI) 
-for Australia and its capital cities which allows for adjustment of the value of Australia dollars for inflation. 
+for Australia and its capital cities which allows for adjustment of the value of Australian dollars for inflation. 
 `ausdex` makes these data available with an inflation calculator
-in a convenient Python package with simple programatic and command-line interfaces.
+in a convenient Python package with simple programmatic and command-line interfaces.
 
 # Statement of need
 
-Understanding price inflation is both critical to any economic analysis, and is one of the key economic concerns in a post-pandemic global economy [@hung2021output]. Therefore it is important to be able to quickly incorporate measures of inflation, and correct historic prices to current values in order to reflect the real change in purchasing power, and the value of goods. To that end we have built a package to automate the collection and calculation of inflation data for Australia, as well as for specific regions of Australia, using data supplied by the ABS.
+Understanding price inflation is critical to any economic analysis and is one of the key economic concerns in a post-pandemic global economy [@hung2021output]. It is important to be able to quickly incorporate measures of inflation and correct historic prices to current values in order to reflect the real change in purchasing power and the value of goods. To that end, we have built a package to automate the collection and calculation of inflation data for Australia, as well as for major Australian cities, using data supplied by the ABS.
 
-The ABS data is generally housed in Microsoft Excel spreadsheets linked from the data catalogue. Working with these spreadsheets directly is cumbersome. This package provides an Application Programming Interface (API) for Australian CPI data that seemlessly interoperates with `NumPy` [@harris2020array] and `pandas` [@mckinney-proc-scipy-2010]. It makes working with Australian dollars in Python convenient in a similar manner to the [cpi](https://github.com/palewire/cpi) Python package which adjusts US dollars for inflation.
+ABS datasets are generally housed in Microsoft Excel spreadsheets linked from the data catalogue. Working with these spreadsheets directly is cumbersome. The `ausdex` package provides an Application Programming Interface (API) for Australian CPI data that seemlessly interoperates with `NumPy` [@harris2020array] and `pandas` [@mckinney-proc-scipy-2010]. It makes working with Australian dollars in Python convenient in a similar manner to the [cpi](https://github.com/palewire/cpi) Python package which adjusts US dollars for inflation.
 
-One example of an application of this package is during an analysis of a time series of building permits in Victoria Australia. corrections to inflation were needed in order to compare the costs of renovations and new builds along a multidecadal dataset. The results of these studies are being presented at the world building conference [cite the world building conference?], and will be included in forthcoming publications [make cite a future vidal and robert paper in prep?]
+This software project was developed for the purpose of performing an analysis of a time series of building permits in Victoria, Australia. Corrections to inflation were needed in order to compare the costs of renovations and new building project along a multi-decadal timespan. The results of these studies are being presented at the 2022 CIB World Building Conference [@crawford_trends_2022; @paton-cole_trends_2022]. This software would be of use for any analysis of Australian prices over a period of time where inflationary effects need to be accounted for.
 
 # Calculating Inflation
 
[…]
