+++ paper.bib
@@ -126,4 +126,19 @@ numpages = {14}
 }
 
 
+@inproceedings{crawford_trends_2022,
+	title = {Trends in {Residential} {Sustainability} {Measures} in the {State} of {Victoria}},
+	author = {Crawford, Robert and Paton-Cole, Vidal and Turnbull, Robert and Fitzgerald, Emily and Garber, Jonathan and Michalewicz, Aleksandra},
+  booktitle = { CIB World Building Congress },
+	year = {2022},
+	pages = {forthcoming},
+}
 
+
+@inproceedings{paton-cole_trends_2022,
+	title = {Trends in {Residential} {Building} {Materials} in the {State} of {Victoria}},
+	author = {Paton-Cole, Vidal and Crawford, Robert and Turnbull, Robert and Fitzgerald, Emily and Garber, Jonathan and Michalewicz, Aleksandra},
+  booktitle = { CIB World Building Congress },
+	year = {2022},
+	pages = {forthcoming},
+}
