### 2022-05-19 8b3d3ac
:tv: trying to get coverage and docs to be deployed only on main and with python 3.10
