+++ .github/workflows/docs.yml
@@ -1,23 +1,0 @@
-name: docs
-on: [push]
-jobs:
-  build-and-deploy:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout 🛎️
-        uses: actions/checkout@v2.3.1
-
-      - name: Install and Build 🔧
-        run: |
-          sudo apt-get install pandoc
-          python -m pip install --upgrade pip
-          pip install poetry
-          poetry install
-          poetry run sphinx-build -b html docs docs/_build/html
-
-      - name: Deploy 🚀
-        uses: JamesIves/github-pages-deploy-action@4.1.5
-        with:
-          branch: gh-pages # The branch the action should deploy to.
-          folder: docs/_build/html # The folder the action should deploy.
-
