+++ .github/workflows/pipeline.yml
@@ -23,13 +23,25 @@ jobs:
         python -m pip install --upgrade pip
         pip install poetry
         poetry install
-    - name: Coverage
+    - name: Testing
       run: |
         poetry run coverage run -m pytest
-        poetry run coverage html
+    - name: Docs
+      run: |
+        poetry run sphinx-build -b html docs gh-pages
+    - name: Coverage
+      run: |
         echo "COVERAGE=$(poetry run coverage report --precision 2 | grep TOTAL | tr -s ' ' | cut -f 4 -d " ")" >> $GITHUB_ENV
-    - name: Create Badge
+        poetry run coverage html --directory gh-pages/coverage
+    - name: Deploy 🚀
+      uses: JamesIves/github-pages-deploy-action@4.1.5
+      if: success() && github.ref == 'refs/heads/main' && ${{ matrix.python-version }} == "3.10"
+      with:
+        branch: gh-pages # The branch the action should deploy to.
+        folder: gh-pages # The folder the action should deploy.
+    - name: Create Coverage Badge
       uses: schneegans/dynamic-badges-action@v1.1.0
+      if: success() && github.ref == 'refs/heads/main' && ${{ matrix.python-version }} == "3.10"
       with:
         auth: ${{ secrets.GIST_SECRET }}
         gistID: 49262550cc8b0fb671d46df58de213d4
