### 2022-05-19 a48626e
:tv: adding pandoc to pipeline
