+++ .github/workflows/pipeline.yml
@@ -20,6 +20,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
+        sudo apt-get install pandoc
         python -m pip install --upgrade pip
         pip install poetry
         poetry install
