### 2022-05-19 c914766
:tv: refactoring actions so that coverage is only done on the main branch for one python version
