+++ .github/workflows/docs.yml
@@ -1,0 +1,43 @@
+name: docs
+
+on:
+  push:
+    branches: main
+      
+jobs:
+  build-and-deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout 🛎️
+        uses: actions/checkout@v2.3.1
+
+      - name: Dependencies 🔧
+        run: |
+          sudo apt-get install pandoc
+          python -m pip install --upgrade pip
+          pip install poetry
+          poetry install
+      - name: Testing
+        run: |
+          poetry run coverage run -m pytest
+      - name: Docs
+        run: |
+          poetry run sphinx-build -b html docs gh-pages
+      - name: Coverage
+        run: |
+          echo "COVERAGE=$(poetry run coverage report --precision 2 | grep TOTAL | tr -s ' ' | cut -f 4 -d " ")" >> $GITHUB_ENV
+          poetry run coverage html --directory gh-pages/coverage
+      - name: Deploy 🚀
+        uses: JamesIves/github-pages-deploy-action@4.1.5
+        with:
+          branch: gh-pages # The branch the action should deploy to.
+          folder: gh-pages # The folder the action should deploy.
+      - name: Create Coverage Badge
+        uses: schneegans/dynamic-badges-action@v1.1.0
+        with:
+          auth: ${{ secrets.GIST_SECRET }}
+          gistID: 49262550cc8b0fb671d46df58de213d4
+          filename: coverage-badge.json
+          label: coverage
+          message: ${{ env.COVERAGE }}
+          color: green
