+++ .github/workflows/pipeline.yml
@@ -4,7 +4,6 @@ name: pipeline
 on: [push]
 jobs:
   build:
-
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -12,7 +11,6 @@ jobs:
         python-version: ["3.7", "3.8", "3.9", "3.10"]
 
     steps:
-      
     - uses: actions/checkout@v2
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
@@ -20,34 +18,10 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        sudo apt-get install pandoc
         python -m pip install --upgrade pip
         pip install poetry
         poetry install
     - name: Testing
       run: |
-        poetry run coverage run -m pytest
-    - name: Docs
-      run: |
-        poetry run sphinx-build -b html docs gh-pages
-    - name: Coverage
-      run: |
-        echo "COVERAGE=$(poetry run coverage report --precision 2 | grep TOTAL | tr -s ' ' | cut -f 4 -d " ")" >> $GITHUB_ENV
-        poetry run coverage html --directory gh-pages/coverage
-    - name: Deploy 🚀
-      uses: JamesIves/github-pages-deploy-action@4.1.5
-      if: success() && github.ref == 'refs/heads/main' && ${{ matrix.python-version }} == "3.10"
-      with:
-        branch: gh-pages # The branch the action should deploy to.
-        folder: gh-pages # The folder the action should deploy.
-    - name: Create Coverage Badge
-      uses: schneegans/dynamic-badges-action@v1.1.0
-      if: success() && github.ref == 'refs/heads/main' && ${{ matrix.python-version }} == "3.10"
-      with:
-        auth: ${{ secrets.GIST_SECRET }}
-        gistID: 49262550cc8b0fb671d46df58de213d4
-        filename: coverage-badge.json
-        label: coverage
-        message: ${{ env.COVERAGE }}
-        color: green
+        poetry run pytest
 
