### 2022-05-19 49ae9b3
:abc: fixing issue with displaying figures in the docs in a notebook
