+++ docs/index.rst
@@ -23,7 +23,7 @@ available in a convenient Python package with a simple programatic and command l
    quickstart
    reference
    cli
-   data_viz   
+   visualization   
    citation
    contributing
    
