+++ .github/workflows/joss-draft-pdf.yml
@@ -1,3 +1,5 @@
+name: joss-draft-pdf
+
 on: [push]
 
 jobs:
