+++ _sources/visualization.ipynb.txt
@@ -5,13 +5,9 @@
    "id": "46f954c6",
    "metadata": {},
    "source": [
-    "# Data Visualization\n",
+    "# Visualization\n",
     "\n",
-    "the ausdex module comes with several data visulisations out of the box build with [plotly](https://plotly.com/python/).\n",
-    "\n",
-    "Visualizations can be created via the command line and saved as html files, or created programmatically.\n",
-    "\n",
-    "Inflation data can be visulised by a time series of the values of a given price inflated to a particular \"dollar year\". Additionally, the consumer price index (CPI) time series can be plotted."
+    "The `ausdex` module comes with several data visulizations built with [plotly](https://plotly.com/python/). Inflation data can be visulized by a time series of the values of a given price inflated to a particular \"dollar year\". Additionally, the consumer price index (CPI) time series can be plotted. Visualizations can be created programmatically or via the command line and saved as images or HTML."
    ]
   },
   {
@@ -36,9 +32,8 @@
    "metadata": {},
    "source": [
     "## Consumer Price Index (CPI) over time\n",
-    "The time series of the CPI can be created as an html file via the command line or programmatically via the API.\n",
     "\n",
-    "Using the following commands, a time series fo the CPI can be generated via the following command:\n",
+    "A time series fo the CPI can be generated via the following command:\n",
     "\n",
     "```bash\n",
     "ausdex plot-cpi [OPTIONS]\n",
@@ -59,7 +54,7 @@
     "- `--title`                     A custom title of the plot.\n",
     "- `--help`                          Show the help message.\n",
     "  \n",
-    "The same plot can be made programmatically via the API. More information found in the [API docs](reference.md):"
+    "The same plot can be made programmatically via the API. More information found in the [API docs](reference.md). See the following example:"
    ]
   },
   {
@@ -157,9 +152,9 @@
     {
      "data": {
       "text/html": [
[…]
