+++ README.md
@@ -7,14 +7,15 @@
 [![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg)](code_of_conduct.md)
 [![status](https://joss.theoj.org/papers/817baa72d2b17b535af8f421a43404b0/status.svg)](https://joss.theoj.org/papers/817baa72d2b17b535af8f421a43404b0)
 
-An interface for several Australian socio-economic indexes.
+A Python package for adjusting Australian dollars for inflation.
 
-The Australian Bureau of Statistics (ABS) publishes a variety of indexes for the Australian
-economic environment. These include the Consumer Price Index (CPI) used for calculating inflation
-and a variety of indexes designed to measure socio-economic advantage. `ausdex` makes these data
-available in a convenient Python package with a simple programatic and command line interfaces.
+The Australian Bureau of Statistics (ABS) publishes the Consumer Price Index (CPI) 
+for Australia and its capital cities which allows for adjustment of the value of Australian dollars for inflation. 
+`ausdex` makes these data available with an inflation calculator in a convenient Python package with simple programmatic and command-line interfaces.
 
-The following information will help you get started with `ausdex`. Further documentation is found here: https://rbturnbull.github.io/ausdex
+ABS datasets are generally housed in Microsoft Excel spreadsheets linked from the data catalogue. Working with these spreadsheets directly is cumbersome. The `ausdex` package provides an Application Programming Interface (API) for Australian CPI data that seemlessly interoperates with `NumPy` and `pandas`. It makes working with Australian dollars in Python convenient in a similar manner to the [cpi](https://github.com/palewire/cpi) Python package which adjusts US dollars for inflation.
+
+The package is documented here: https://rbturnbull.github.io/ausdex
 
 ## Installation
 
@@ -23,6 +24,8 @@ You can install `ausdex` from the Python Package Index (PyPI):
 ```
 pip install ausdex
 ```
+
+To install ausdex for development, see the documentation for [contributing](https://rbturnbull.github.io/ausdex/contributing.html).
 
 ## Command Line Usage
 
@@ -88,14 +91,14 @@ See the guidelines for contributing and our code of conduct in the [documentatio
 
 ## License and Disclaimer
 
-ausdex is released under the Apache 2.0 license.
+`ausdex` is released under the Apache 2.0 license.
 
 While every effort has been made by the authors of this package to ensure that the data and calculations used to produce the results are accurate, as is stated in the license, we accept no liability or responsibility for the accuracy or completeness of the calculations. 
 We recommend that users exercise their own care and judgment with respect to the use of this package.
  
 ## Credits
 
-ausdex was written by [Dr Robert Turnbull](https://findanexpert.unimelb.edu.au/profile/877006-robert-turnbull) and [Dr Jonathan Garber](https://findanexpert.unimelb.edu.au/profile/787135-jonathan-garber) from the [Melbourne Data Analytics Platform](https://mdap.unimelb.edu.au/).
+`ausdex` was written by [Dr Robert Turnbull](https://findanexpert.unimelb.edu.au/profile/877006-robert-turnbull) and [Dr Jonathan Garber](https://findanexpert.unimelb.edu.au/profile/787135-jonathan-garber) from the [Melbourne Data Analytics Platform](https://mdap.unimelb.edu.au/).
 
 Please cite from the article when it is released. Details to come soon.
 
