### 2022-05-19 #4212 @rbturnbull
Hi @ajstewartlang and @konstantinstadler - we have put the SEIFA code in another repository (https://github.com/sailngarbwm/seifa/). The issues that were relevant to that module were moved there. I've added in some new functionality to ausdex so that it can calculate the inflation for particular capital cities in Australia. We've updated the docs quite a bit and edited the text of the paper in light of the feedback we received. We published a new release to PyPI today (v1.0.1). Hopefully we've covered everything. Please let us know if there are any further things we should address.
