### 2022-05-19 #4212 @konstantinstadler
That seems ok for me. I tested the new version and it runs without problems. Some questions to @ajstewartlang :

- I guess the article needs to be rebuild to incorporate the changes. Not sure how to initiate that
- Just to double-check: for this release we now only consider ausdex - seifa becomes a separate article
