### 2022-05-19 #4212 @ajstewartlang
Many thanks, yes, @konstantinstadler - the article will need to be rewritten to incorporate the changes and new focus. SEIFA will now be in a separate repo (and outside the scope of the present submission) - the article will need to be updated to reflect that.
