### 2022-05-19 #4212 @konstantinstadler
Looks good to me!
It is much more focused then before - definitely a useful package.
All issues connected to the ausdex module have been resolved - I recommend to publish it.
