### 2022-06-13 rbturnbull/ausdex#7 @potterzot
Add additional detail to the installation section of documentation
The installation currently does not note that `ausdex` only works with Python versions `>=3.7, <3.10`. A sentence or two about these limitations would be helpful.

As an example, I am working from a newer laptop that only has 3.10 installed, and installation errors were not indicative of these version limitations. I had to try the poetry installation to find these version limitations, and then install python 3.7.
