### 2022-06-13 rbturnbull/ausdex#8 @potterzot
Update package to work with python 3.10
How difficult would this be? Not necessary for the review but potentially helpful for future compatibility.
