### 2022-06-13 rbturnbull/ausdex#9 @potterzot
Issues with the usage documentation
Some small suggestions for the usage documentation:

- The example for adjusting values says "today", but with no indication of the frequency of CPI data releases. A user familiar with the ABS CPI data would know this, but a user (e.g. myself) who is not would not know that it is quarterly. I would remove the use of "today" throughout and use something like "most recent quarter".
- is "evaluation-date" standard terminology? I would suggest modifying the documentation to make it clear that the evaluation date means the outcome is reported in real dollars at that date.
