### 2022-06-13 rbturnbull/ausdex#10 @potterzot
Add functionality to allow user download over a range of dates and to a specific location
The package suggests that it provides user access to data, and it does, but a key use for my workflow would be to download the raw data itself, because I want to have access to those numbers if I need to replicate something.

It looks like currently `ausdex.cached_download()` is the option to download, but a user-specific method that allows for downloading the CPI data over a range of dates and into a specific folder would be useful. Adding this to the usage documentation would also be helpful.
