### 2022-06-13 #4212 @potterzot
Review checklist for @potterzot

----

My apologies on taking a substantial amount of time for this. I could offer a number of reasons but it really doesn't justify the time that I've taken. Reviewers should be prompt and I have not been. I hope the following review will be helpful. There are several small suggestions that I have listed below and created issues for in the project repository. Most have to do with the documentation, but there is also one additional method suggestion. Overall, I think this is a worthwhile package and it is helpful to have a way to programmatically access important data such as CPI data.

I am leaving the `installation instructions` and `example usage` boxes unchecked pending resolution of the issues below.

## Review

Thank you to @konstantinstadler for your review, which helped me as I was reviewing this package.

The package `ausdex` provides access to CPI data from the Australian Bureau of Statistics. It allows for downloading CPI data and using that data to adjust to real value for a given date. The package is valuable since it provides programmatic access to CPI data via python in a way that makes research more reproducible and helps to automate sequential downloads for web apps or other uses.

In general, while the package documentation covers the basics of installation and use, it does not provide details that would be useful to users. Installation documentation is limited to a single line, but fails to note the Python version restrictions for the package or suggest ways to proceed if installation fails. The documentation for uses provides a few examples, but does not explain the organizations of the module in a way that allows the user to see the full scope of the module functionality. Documenting the full suite of capabilities would improve the ease of use substantially.

It is also unclear from the documentation whether the package can be used to download the CPI, rather than just use to adjust. From the package description and from looking at the code, it seems it certainly can, but that is not detailed in the usage section of the documentation. For me, this would be the most valuable aspect of this package, since I would want the CPI as a data frame that I can then use to perform my own calculations on prices.

### Specific issues are filed as issues on the repository, but listed here for reference
- [Issue 7](https://github.com/rbturnbull/ausdex/issues/7): Given that the goal of the package is to make data easily available, I would suggest expanding the installation section of the README. In particular, the package currently does not work with Python 3.10. It would be worth noting that in the installation, and suggesting that the user install Python 3.7 specifically.
- [Issue 8](https://github.com/rbturnbull/ausdex/issues/8): Speaking of, how difficult would it be to update the package to work with Python 3.10? While not strictly necessary, it seems like it would be helpful going forward.
- [Issue 9](https://github.com/rbturnbull/ausdex/issues/9): Updates to documentation. The example for adjusting values says "today", but with no indication of the frequency of CPI data releases. A user familiar with the ABS CPI data would know, but a user (e.g. myself) who is not would not know that it is quarterly. I would remove the use of "today" throughout and use something like "most recent quarter". Also, is "evaluation-date" standard terminology? I would suggest modifying the documentation to make it clear that the evaluation date means the outcome is reported in real dollars at that date.
- [Issue 10](https://github.com/rbturnbull/ausdex/issues/10): Add functionality and documentation to detail how to download the CPI data directly. It looks like currently `ausdex.cached_download()` is the option to download, but a user-specific method that allows for downloading the CPI data over a range of dates and into a specific folder would be useful.

----
## Checklist

- [X] As the reviewer I confirm that I have read the [JOSS conflict of interest policy](reviewer_guidelines.html#joss-conflict-of-interest-policy) and that there are no conflicts of interest for me to review this work.

- [X] **Repository:** Is the source code for this software available at the [repository url](https://github.com/rbturnbull/ausdex)?
- [X] **Contribution and Authorship:** Has the submitting author made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
