### 2022-06-13 #4212 @rbturnbull
Thanks @potterzot. That's all great feedback. I'll get to work on those issues right away
