### 2022-06-13 rbturnbull/ausdex#7 @rbturnbull
I'll make it so that it is compatible with python 3.10 as per issue #8. When we started the package last year, one of the dependencies required python<3.10 but now that is no longer required.
