### 2022-06-13 3632fa3
:tv: removing requirement of python<3.10. This addresses issues #7 and #8. These issues can be closed after the next release to PyPI.
