+++ pyproject.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/rbturnbull/ausdex"
 documentation = "https://rbturnbull.github.io/ausdex/"
 
 [tool.poetry.dependencies]
-python = "^3.7.1,<3.10"
+python = "^3.7.1"
 appdirs = "^1.4.4"
 pandas = "^1.3.2"
 openpyxl = "^3.0.0"
