### 2022-06-14 1108556
:zap: refactoring code to download and cache files so that users can save files to specified locations if desired. For issue #10
