+++ ausdex/inflation.py
@@ -2,9 +2,8 @@ import sys
 from enum import Enum
 from pathlib import Path
 from datetime import datetime, timedelta
-from typing import Union, List
+from typing import Union, List, Optional
 import pandas as pd
-from pandas._config import config
 import modin.pandas as mpd
 import numpy as np
 import numbers
@@ -13,7 +12,7 @@ import plotly.express as px
 
 from cached_property import cached_property
 
-from .files import cached_download, get_cached_path, DownloadError
+from .files import latest_cpi_datafile
 from .dates import convert_date
 from .viz import format_fig
 
@@ -36,79 +35,6 @@ class Location(str, Enum):
 class CPI:
     """A class to manage the Australian Consumer Index (CPI) data."""
 
-    ACCEPTED_QUARTERS = ["mar", "jun", "sep", "dec"]
-
-    def get_abs(self, id: str, quarter: str, year: Union[str, int]) -> Path:
-        """Gets a datafile from the Australian Burau of Statistics.
-
-        Args:
-            id (str): The ABS id for the datafile. For Australian Consumer Price Index the ID is 640101.
-            quarter (str): One of "mar", "jun", "sep", or "dec".
-            year (Union[str,int]): [description]
-
-        Raises:
-            ValueError: Raises this error if the quarter cannot be understood.
-
-        Returns:
-            Path: The path to the cached ABS datafile
-        """
-        quarter = quarter.lower()[:3]
-        if quarter not in self.ACCEPTED_QUARTERS:
-            raise ValueError(f"Cannot understand quarter {quarter}.")
-
-        def cached_download_abs(quarter, year, id, extension):
-            url = f"https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia/{quarter}-{year}/{id}.{extension}"
-            local_path = get_cached_path(f"{id}-{quarter}-{year}.{extension}")
-            cached_download(url, local_path)
-            return local_path
-
-        try:
-            local_path = cached_download_abs(quarter=quarter, year=year, id=id, extension="xlsx")
-        except (DownloadError, IOError):
-            local_path = cached_download_abs(quarter=quarter, year=year, id=id, extension="xls")
-
-        return local_path
-
-    def get_abs_by_date(self, id: str, date: datetime) -> Path:
-        """
-        Gets a datafile from the Australian Burau of Statistics before a specific date.
-
-        Args:
-            id (str): The ABS id for the datafile. For Australian Consumer Price Index (CPI) the ID is 640101.
-            date (datetime): The date before which the CPI data should be valid.
-
-        Returns:
-            Path: The path to the cached ABS datafile.
-        """
-        file = None
-        while file is None and date > datetime(1948, 1, 1):
-            try:
-                year = date.year
-                quarter_index = (date.month - 3) // 3
-                if quarter_index == -1:
-                    quarter_index = 3
-                    year -= 1
-                quarter = self.ACCEPTED_QUARTERS[quarter_index]
-
-                file = self.get_abs(id, quarter, year)
-            except:
-                print(f"CPI data for {date} not available.", file=sys.stderr)
-
-            date -= timedelta(days=89)  # go back approximately a quarter
-
-        return file
-
-    def latest_cpi_datafile(self) -> Path:
-        """
-        Returns the path to the latest cached file with the Australian Consumer Price Index (CPI) data.
-
-        The ABS id of this file is "640101".
-
-        Returns:
-            Path: The path to the cached datafile.
-        """
-        return self.get_abs_by_date("640101", datetime.now())
-
     @cached_property
     def latest_cpi_df(self) -> pd.DataFrame:
         """
@@ -117,7 +43,7 @@ class CPI:
         Returns:
[…]
