+++ ausdex/files.py
@@ -1,33 +1,46 @@
+import sys
 from pathlib import Path
 from appdirs import user_cache_dir
-
+from typing import Union
 import urllib.request
+from datetime import datetime, timedelta
 
 
-def get_cached_path(filename):
-    cache_dir = Path(user_cache_dir("ausdex"))
-    cache_dir.mkdir(exist_ok=True, parents=True)
-    return cache_dir / filename
+ACCEPTED_QUARTERS = ("mar", "jun", "sep", "dec")
 
 
 class DownloadError(Exception):
     pass
 
 
-def cached_download(url: str, local_path: (str, Path), attempt_download=True) -> None:
+def get_cached_path(filename: str) -> Path:
+    """
+    Returns a path in the ausdex directory in the user's cache.
+
+    File may or may not exist.
+    """
+    cache_dir = Path(user_cache_dir("ausdex"))
+    cache_dir.mkdir(exist_ok=True, parents=True)
+    return cache_dir / filename
+
+
+def cached_download(url: str, local_path: Union[str, Path], force: bool = False) -> None:
     """
     Downloads a file if a local file does not already exist.
 
     Args:
-        url: The url of the file to download.
-        local_path: The local path of where the file should be. If this file isn't there or the file size is zero then this function downloads it to this location.
+        url (str): The url of the file to download.
+        local_path (str, Path): The local path of where the file should be.
+            If this file isn't there or the file size is zero then this function downloads it to this location.
+        force (bool): Whether or not the file should be forced to download again even if present in the local path.
+            Default False.
 
     Raises:
-        Exception: Raises an exception if it cannot download the file.
-
+        DownloadError: Raises an exception if it cannot download the file.
+        IOError: Raises an exception if the file does not exist or is empty after downloading.
     """
     local_path = Path(local_path)
-    if (not local_path.exists() or local_path.stat().st_size == 0) and attempt_download:
+    if (not local_path.exists() or local_path.stat().st_size == 0) or force:
         try:
             print(f"Downloading {url} to {local_path}")
             urllib.request.urlretrieve(url, local_path)
@@ -36,3 +49,129 @@ def cached_download(url: str, local_path: (str, Path), attempt_download=True) ->
 
     if not local_path.exists() or local_path.stat().st_size == 0:
         raise IOError(f"Error reading {local_path}")
+
+
+def cached_download_abs(
+    id: str,
+    quarter: str,
+    year: Union[int, str],
+    extension: str,
+    local_path: Union[Path, str, None] = None,
+    force: bool = False,
+) -> Path:
+    """
+    Downloads a file from the ABS if a local file does not already exist.
+
+    Args:
+        id (str): The ABS id for the datafile. For Australian Consumer Price Index the ID is 640101.
+        quarter (str): The quarter of the file in question. One of "mar", "jun", "sep", or "dec".
+        year (str, int): The year for the file in question.
+        extension (str): The extension of the file in question.
+        local_path (Path, str, optional): The path to where the file should be downloaded.
+            If None, then it is downloaded in the user's cache directory.
+        force (bool): Whether or not the file should be forced to download again even if present in the local path.
+        local_path (str, Path): The local path of where the file should be.
+            If this file isn't there or the file size is zero then this function downloads it to this location.
+        force (bool): Whether or not the file should be forced to download again even if present in the local path.
+
+    Raises:
+        ValueError: If the value for `quarter` cannot be understood.
+        DownloadError: Raises an exception if it cannot download the file.
+        IOError: Raises an exception if the file does not exist or is empty after downloading.
+
+    Returns:
[…]
