+++ docs/reference.rst
@@ -2,7 +2,16 @@
 API Reference
 =======================
 
+Inflation 
+======================
+
 .. automodule:: ausdex.inflation
+   :members:   
+
+Files 
+======================
+
+.. automodule:: ausdex.files
    :members:   
 
 Main 
