### 2022-06-14 5528ad7
:tv: updating numpy to >=1.22.0 which allows python 3.10. However it requires python >= 3.8. Updating the installation section in the docs with this new requirement. Addresses issues #7 and #8.
