+++ pyproject.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/rbturnbull/ausdex"
 documentation = "https://rbturnbull.github.io/ausdex/"
 
 [tool.poetry.dependencies]
-python = "^3.7.1"
+python = ">=3.8,<4.0.0"
 appdirs = "^1.4.4"
 pandas = "^1.3.2"
 openpyxl = "^3.0.0"
@@ -21,6 +21,7 @@ modin = "^0.10.2"
 importlib-metadata = "<4.3"
 plotly = "^5.4.0"
 kaleido = "0.2.1"
+numpy = "^1.22.0"
 
 [tool.poetry.dev-dependencies]
 ipykernel = "^6.2.0"
