### 2022-06-14 rbturnbull/ausdex#9 @rbturnbull
Good feedback. I'll change the documentation as you suggested.
