### 2022-06-14 b24b6bb
:abc: updating documentation to explain more clearly what 'evaluation_date' is and that if it is not given then it uses the date of the most recent quarter. Addresses issue #9
