### 2022-06-14 fd4f61e
:zap: renaming file to download CPI excel file to be clearer and allowing you to specify the date explicitly
