### 2022-06-14 rbturnbull/ausdex#10 @rbturnbull
Yeah, that's a much better idea. I've refactored to put all the downloading code into files.py. There is now a function `cached_download_cpi` which allows the user to download to a specified local path if given. You can also set the date explicitly. This function and all others in files.py are now added to the API section of the docs here: https://rbturnbull.github.io/ausdex/reference.html#module-ausdex.files
