+++ .github/workflows/docs.yml
@@ -5,39 +5,43 @@ on:
     branches: main
       
 jobs:
-  build-and-deploy:
+  build:
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.9"]
     steps:
-      - name: Checkout 🛎️
-        uses: actions/checkout@v2.3.1
-
-      - name: Dependencies 🔧
-        run: |
-          sudo apt-get install pandoc
-          python -m pip install --upgrade pip
-          pip install poetry
-          poetry install
-      - name: Testing
-        run: |
-          poetry run coverage run -m pytest
-      - name: Docs
-        run: |
-          poetry run sphinx-build -b html docs gh-pages
-      - name: Coverage
-        run: |
-          echo "COVERAGE=$(poetry run coverage report --precision 2 | grep TOTAL | tr -s ' ' | cut -f 4 -d " ")" >> $GITHUB_ENV
-          poetry run coverage html --directory gh-pages/coverage
-      - name: Deploy 🚀
-        uses: JamesIves/github-pages-deploy-action@4.1.5
-        with:
-          branch: gh-pages # The branch the action should deploy to.
-          folder: gh-pages # The folder the action should deploy.
-      - name: Create Coverage Badge
-        uses: schneegans/dynamic-badges-action@v1.1.0
-        with:
-          auth: ${{ secrets.GIST_SECRET }}
-          gistID: 49262550cc8b0fb671d46df58de213d4
-          filename: coverage-badge.json
-          label: coverage
-          message: ${{ env.COVERAGE }}
-          color: green
+    - uses: actions/checkout@v3
+    - name: Install poetry
+      run: pipx install poetry
+    - name: Install dependencies for Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v3
+      with:
+        python-version: ${{ matrix.python-version }}
+        cache: 'poetry'
+    - run: poetry install
+    - name: Docs
+      run: |
+        poetry run sphinx-build -b html docs gh-pages
+    - name: Coverage
+      run: |
+        poetry run coverage run -m pytest
+        echo "COVERAGE=$(poetry run coverage report --precision 2 | grep TOTAL | tr -s ' ' | cut -f 4 -d " ")" >> $GITHUB_ENV
+        poetry run coverage html --directory gh-pages/coverage
+    - name: Deploy 🚀
+      uses: JamesIves/github-pages-deploy-action@4.1.5
+      with:
+        branch: gh-pages # The branch the action should deploy to.
+        folder: gh-pages # The folder the action should deploy.
+    
+    - name: Create Coverage Badge
+      uses: schneegans/dynamic-badges-action@v1.1.0
+      with:
+        auth: ${{ secrets.GIST_SECRET }}
+        gistID: 49262550cc8b0fb671d46df58de213d4
+        filename: coverage-badge.json
+        label: coverage
+        message: ${{ env.COVERAGE }}
+        color: green
+    
