+++ .github/workflows/pipeline.yml
@@ -1,4 +1,3 @@
-# Based on https://github.com/actions/starter-workflows/blob/main/ci/python-package.yml
 name: pipeline
 
 on: [push]
@@ -8,20 +7,17 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.7", "3.8", "3.9", "3.10"]
-
+        python-version: ["3.8", "3.9", "3.10"]
     steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
+    - uses: actions/checkout@v3
+    - name: Install poetry
+      run: pipx install poetry
+    - name: Install dependencies for Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v3
       with:
         python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install poetry
-        poetry install
+        cache: 'poetry'
+    - run: poetry install
     - name: Testing
       run: |
         poetry run pytest
-
