### 2022-06-14 9b624fd
:tv: adding pandoc to dependencies for docs CI
