+++ .github/workflows/docs.yml
@@ -14,7 +14,9 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install poetry
-      run: pipx install poetry
+      run: |
+        sudo apt-get install pandoc
+        pipx install poetry
     - name: Install dependencies for Python ${{ matrix.python-version }}
       uses: actions/setup-python@v3
       with:
