### 2022-06-14 a90d953
:tv: bumping version number to prepare for release to pypi
