+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "ausdex"
-version = "1.0.1"
+version = "1.1.0"
 description = "A Python package for adjusting Australian dollars for inflation"
 authors = ["Robert Turnbull <robert.turnbull@unimelb.edu.au>", "Jonathan Garber <jonathan.garber@unimelb.edu.au>"]
 license = "Apache-2.0"
