### 2022-06-14 rbturnbull/ausdex#8 @rbturnbull
Closing because the new version is compatible with python 3.10 and it has been released to PyPI.
