### 2022-06-14 rbturnbull/ausdex#7 @rbturnbull
Closing because the new version is compatible with python 3.10 and it has been released to PyPI. There is a note in the documentation that it needs python >=3.8.
