+++ README.md
@@ -25,6 +25,8 @@ You can install `ausdex` from the Python Package Index (PyPI):
 pip install ausdex
 ```
 
+`ausdex` requires Python 3.8 or higher.
+
 To install ausdex for development, see the documentation for [contributing](https://rbturnbull.github.io/ausdex/contributing.html).
 
 ## Command Line Usage
