### 2022-06-14 fa52f6b
:zap: adding functionality to get the dataframe of the raw CPI data. adding information about it to the docs
