+++ ausdex/__init__.py
@@ -1,1 +1,1 @@
-from .inflation import calc_inflation, Location
+from .inflation import calc_inflation, Location, latest_cpi_df
