### 2022-06-14 #4212 @rbturnbull
Hi @potterzot - I've gone through the issues and issued a new release to PyPI. In addition to the issues, I added a function to easily get the raw CPI data as a pandas dataframe. I've added information about how to use this function in the docs. Please let me know if I need to do anything else.
