### 2022-07-05 #4212 @rbturnbull
Hi @ajstewartlang  - I created a release (v1.1.1) and published it on figshare: https://melbourne.figshare.com/articles/software/ausdex_A_Python_package_for_adjusting_Australian_dollars_for_inflation/20226120
Here is the DOI: https://doi.org/10.26188/20226120.v1
Please let me know if there is anything else needed.
