### 2022-07-05 #4212 @ajstewartlang
> Hi @ajstewartlang - the DOI link from the editorial includes doi.org twice so the link won't work

should be fixed now
