### 2022-07-05 #4212 @danielskatz
@rbturnbull - I'm the AEiC on duty this week, and will finish the processing of your paper.  I've suggested some minor changes to the paper in https://github.com/rbturnbull/ausdex/pull/13 - please merge this or let me know what you disagree with, then we can proceed.
