### 2022-07-26 35180fa
:zap: refactoring visualization code and adding new function to visualize CPI increase over time
