### 2022-07-27 ef59f8f
:zap: adding more options to plot_cpi_change
