### 2022-10-11 #4212 @ajstewartlang
Hi - yes, that's correct. I believe they've been on vacation the last
couple of weeks but will get to your submission asap.

On Sun, 29 May 2022, 04:01 Robert Turnbull, ***@***.***>
wrote:

> Hi @ajstewartlang <https://github.com/ajstewartlang> - am I right that
> the next step is to hear back from the other reviewer?
>
